에러발생 리팩토링중

+++ backend/src/main/java/com/example/backend/travelPackage/dto/PopularPackageDto.java
@@ -17,6 +17,6 @@ public class PopularPackageDto {
     private Long id;
     private String title;
     private String cityName;
-    private Double price;
+    private BigDecimal price;
     private String imageUrl;
 }
